api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-ollama==0.3.6
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.9
MarkupSafe==3.0.2
ollama==0.5.1
openai==1.98.0
orjson==3.11.1
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.42
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
zstandard==0.23.0
